Fix types in CRC16 function

+++ modlib.c
@@ -22,7 +22,7 @@ uint16_t MODBUSSwapEndian( uint16_t Data )
     return Conversion.Data;
 }
 
-uint16_t MODBUSCRC16( uint16_t *Data, uint16_t Length )
+uint16_t MODBUSCRC16( uint8_t *Data, uint16_t Length )
 {
 	//Calculate CRC16 checksum using given data and length
